Fix RefineGAN SineGenerator dtype mismatch on fp16 inference (CUDA)
SineGenerator.forward always produces float32 sine_waves — the source
buffer and all intermediate ops (torch.zeros, cumsum, sin) are computed
at float32 regardless of model precision. The downstream merge Linear
layer is fp16 when net_g.half() is called for CUDA inference, causing:
  RuntimeError: expected mat1 and mat2 to have same dtype: float != c10::Half

Fix: cast sine_waves to the merge layer's weight dtype immediately before
self.merge(). Uses self.merge[0].weight.dtype (direct attribute, no
iteration cost) — same pattern as NSF-HiFi-GAN SourceModuleHnNSF which
calls sine_waves.to(x.dtype) for the same reason.

fp32 models (MPS, CPU) are unaffected — .to(torch.float32) on a float32
tensor is a no-op.

+++ backend/rvc/rvc/layers/refinegan.py
@@ -171,7 +171,12 @@ class SineGenerator(nn.Module):
             noise_amp = uv * self.noise_std + (1 - uv) * self.sine_amp / 3
             noise = noise_amp * torch.randn_like(sine_waves)
             sine_waves = sine_waves * uv + noise
-        return self.merge(sine_waves)
+        # Cast to the model's dtype before entering the learnable merge layer.
+        # sine_waves is always float32 (generated from float32 f0 cache) but
+        # the Linear weight is fp16 when the model runs in half precision.
+        # NSF-HiFi-GAN handles this with an explicit .to(x.dtype); we do the
+        # same here.
+        return self.merge(sine_waves.to(self.merge[0].weight.dtype))
 
 
 # ---------------------------------------------------------------------------
